Convert hair, skin, eye palette coordinates

+++ ImperatorToCK3.UnitTests/Imperator/Characters/CharacterTests.cs
@@ -149,12 +149,12 @@ namespace ImperatorToCK3.UnitTests.Imperator.Characters {
 			);
 			var character = ImperatorToCK3.Imperator.Characters.Character.Parse(reader, "42", genesDB);
 
-			Assert.Equal((uint)0, character.PortraitData.HairColorPaletteCoordinates.x);
-			Assert.Equal((uint)0, character.PortraitData.HairColorPaletteCoordinates.y);
-			Assert.Equal((uint)0, character.PortraitData.SkinColorPaletteCoordinates.x);
-			Assert.Equal((uint)0, character.PortraitData.SkinColorPaletteCoordinates.y);
-			Assert.Equal((uint)0, character.PortraitData.EyeColorPaletteCoordinates.x);
-			Assert.Equal((uint)0, character.PortraitData.EyeColorPaletteCoordinates.y);
+			Assert.Equal(0, character.PortraitData.HairColorPaletteCoordinates.x);
+			Assert.Equal(0, character.PortraitData.HairColorPaletteCoordinates.y);
+			Assert.Equal(0, character.PortraitData.SkinColorPaletteCoordinates.x);
+			Assert.Equal(0, character.PortraitData.SkinColorPaletteCoordinates.y);
+			Assert.Equal(0, character.PortraitData.EyeColorPaletteCoordinates.x);
+			Assert.Equal(0, character.PortraitData.EyeColorPaletteCoordinates.y);
 		}
 
 		[Fact]
